separated hamburger menu from nav tab tests
