+++ tests/conf.js
@@ -18,8 +18,9 @@ const baseConfig = {
   seleniumAddress: 'http://localhost:4444/wd/hub',
   //Enabling specs as they are working.
   suites: {
+    e2e:'e2e/navigation/common.specs.js',
     // e2e:'e2e/**/*.js',
-    e2e: ['**/login.spec.js'],
+    //e2e: ['**/login.spec.js'],
     // performance: 'performance/**/*.js'
   },
   framework: 'jasmine2',
@@ -133,7 +134,7 @@ const login = browser => {
   browser.driver.findElement(by.id('login')).click();
   // Login takes some time, so wait until it's done.
   const bootstrappedCheck = () =>
-    element(by.css('.app-root.bootstrapped')).isPresent();
+    element(by.css('.inbox.bootstrapped')).isPresent();
   return browser.driver.wait(
     bootstrappedCheck,
     20 * 1000,
